feat: refactor faker factory tests
