+++ libraries/faker-factory/src/index.ts
@@ -1,3 +1,4 @@
-export * from './faker-factory';
+export { Settings } from './settings.type';
+export { FakerFactory } from './faker-factory';
 export { toPojo, isPojo } from './pojo';
-export { Settings } from './settings';
+export { getJsonSchemas } from './get-json-schemas';
